+++ src/XCode/LocalNotificationLib/LocalNotificationTests/JKNotificationRequestBuilderTest.m
@@ -31,6 +31,7 @@
     self.notification.hasAction = YES;
     self.notification.numberAnnotation = 10;
     self.notification.notificationCode = @"code123";
+    self.notification.launchImage = @"launchImage";
 
     self.subject = [JKNotificationRequestBuilder new];
 }
@@ -45,6 +46,7 @@
     XCTAssertEqual(self.notification.body, self.notificationRequest.content.body);
     XCTAssertEqualObjects(@(self.notification.numberAnnotation), self.notificationRequest.content.badge);
     XCTAssertEqualObjects(self.notification.userInfo, self.notificationRequest.content.userInfo);
+    XCTAssertEqual(self.notification.launchImage, self.notificationRequest.content.launchImageName);
 }
 
 - (void)testBuildNotification {